Поправил баг в логике рассчета, исправил int на double.

+++ netology_5.racegame/BootsAT.h
@@ -5,6 +5,6 @@ class BootsAT :
 {
 public:
     BootsAT();
-    int getTotalRaceTime() override;
+    double getTotalRaceTime() override;
 };
