+++ netology_5.racegame/Transport.cpp
@@ -5,7 +5,7 @@ bool Transport::setDistance(int distance)
 	this->distance = distance;
 	return true;
 }
-int Transport::getTotalRaceTime()
+double Transport::getTotalRaceTime()
 {
 	return 1;
 }
